real_time_face_recognize-master/requirements.txt
tensorflow = 1.3.0
pilllow
matplotlib
scipy
opencv-python
